ATHENA-306-Criteria builder

+++ src/main/java/com/odysseusinc/athena/model/athenav5/SolrConcept.java
@@ -27,6 +27,7 @@ import java.util.Date;
 import java.util.List;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -44,7 +45,7 @@ public class SolrConcept extends EntityV5 {
     @Id
     @NotBlank
     @Column(name = "concept_id")
-    private Integer id;
+    private Long id;
 
     @NotNull
     @Column(name = "concept_name")
@@ -86,12 +87,12 @@ public class SolrConcept extends EntityV5 {
     @LazyCollection(LazyCollectionOption.FALSE)
     private List<ConceptSynonymV5> synonyms;
 
-    public Integer getId() {
+    public Long getId() {
 
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
 
         this.id = id;
     }
